Remove code duplication from TiltShiftX and Y filters. This resolves a TODO that was in there.
